Main to Dev

+++ auto-install.sh
@@ -1,16 +1,15 @@
 #!/usr/bin/env bash
+set -e
 
 # Set some colors for output messages
 OK="$(tput setaf 2)[OK]$(tput sgr0)"
 ERROR="$(tput setaf 1)[ERROR]$(tput sgr0)"
 NOTE="$(tput setaf 3)[NOTE]$(tput sgr0)"
-WARN="$(tput setaf 166)[WARN]$(tput sgr0)"
+WARN="$(tput setaf 5)[WARN]$(tput sgr0)"
 CAT="$(tput setaf 6)[ACTION]$(tput sgr0)"
 ORANGE=$(tput setaf 166)
 YELLOW=$(tput setaf 3)
 RESET=$(tput sgr0)
-
-initial_dir=$(pwd)
 
 if [ -n "$(grep -i nixos < /etc/os-release)" ]; then
   echo "Verified this is NixOS."
@@ -29,21 +28,8 @@ else
   exit 1
 fi
 
-echo "$NOTE Default options are in brackets []"
-echo "$NOTE Just press enter to select the default"
-sleep 2
-
-echo "-----"
-
 echo "$NOTE Ensure In Home Directory"
 cd || exit
-
-echo "-----"
-
-read -rp "$CAT Enter Your New Hostname: [ default ] " hostName
-if [ -z "$hostName" ]; then
-  hostName="default"
-fi
 
 echo "-----"
 
@@ -52,12 +38,12 @@ if [ -d "NixOS-Hyprland" ]; then
   echo "$NOTE NixOS-Hyprland exists, backing up to NixOS-Hyprland-backups folder."
   if [ -d "NixOS-Hyprland-backups" ]; then
     echo "Moving current version of NixOS-Hyprland to backups folder."
-    mv "$HOME"/NixOS-Hyprland NixOS-Hyprland-backups/"$backupname"
+    sudo mv "$HOME"/NixOS-Hyprland NixOS-Hyprland-backups/"$backupname"
     sleep 1
   else
     echo "$NOTE Creating the backups folder & moving NixOS-Hyprland to it."
     mkdir -p NixOS-Hyprland-backups
-    mv "$HOME"/NixOS-Hyprland NixOS-Hyprland-backups/"$backupname"
+    sudo mv "$HOME"/NixOS-Hyprland NixOS-Hyprland-backups/"$backupname"
     sleep 1
   fi
 else
@@ -67,13 +53,12 @@ fi
 
 echo "-----"
 
-set -e
-
 echo "$NOTE Cloning & Entering NixOS-Hyprland Repository"
 git clone --depth 1 https://github.com/JaKooLit/NixOS-Hyprland.git ~/NixOS-Hyprland
 cd ~/NixOS-Hyprland || exit
 
 printf "\n%.0s" {1..2}
+
 # Checking if running on a VM and enable in default config.nix
 if hostnamectl | grep -q 'Chassis: vm'; then
   echo "${NOTE} Your system is running on a VM. Enabling guest services.."
@@ -86,7 +71,7 @@ printf "\n%.0s" {1..1}
 if command -v lspci > /dev/null 2>&1; then
   # lspci is available, proceed with checking for Nvidia GPU
   if lspci -k | grep -A 2 -E "(VGA|3D)" | grep -iq nvidia; then
-    echo "$NOTE Nvidia GPU detected. Setting up for nvidia..."
+    echo "${NOTE} Nvidia GPU detected. Setting up for nvidia..."
     sed -i '/drivers\.nvidia\.enable = false;/s/drivers\.nvidia\.enable = false;/ drivers.nvidia.enable = true;/' hosts/default/config.nix
   fi
 fi
@@ -94,35 +79,72 @@ fi
 echo "-----"
 printf "\n%.0s" {1..1}
 
-mkdir hosts/"$hostName"
-cp hosts/default/*.nix hosts/"$hostName"
-git config --global user.name "installer"
-git config --global user.email "installer@gmail.com"
-git add .
-sed -i '/^\s*host[[:space:]]*=[[:space:]]*\"[^"]*\"/s/\"\([^"]*\)\"/\"'"$hostName"'\"/' ./flake.nix
+echo "$NOTE Default options are in brackets []"
+echo "$NOTE Just press enter to select the default"
+sleep 1
 
+echo "-----"
+
+read -rp "$CAT Enter Your New Hostname: [ default ] " hostName
+if [ -z "$hostName" ]; then
+  hostName="default"
+fi
+
+echo "-----"
+
+# Create directory for the new hostname, unless the default is selected
+if [ "$hostName" != "default" ]; then
+  mkdir -p hosts/"$hostName"
+  cp hosts/default/*.nix hosts/"$hostName"
+else
+  echo "Default hostname selected, no extra hosts directory created."
+fi
+echo "-----"
 
 read -rp "$CAT Enter your keyboard layout: [ us ] " keyboardLayout
 if [ -z "$keyboardLayout" ]; then
   keyboardLayout="us"
 fi
 
-sed -i "/^\s*keyboardLayout[[:space:]]*=[[:space:]]*\"/s/\"\(.*\)\"/\"$keyboardLayout\"/" ./hosts/$hostName/variables.nix
+sed -i 's/keyboardLayout\s*=\s*"\([^"]*\)"/keyboardLayout = "'"$keyboardLayout"'"/' ./hosts/$hostName/variables.nix
 
 echo "-----"
 
 installusername=$(echo $USER)
-sed -i "/^\s*username[[:space:]]*=[[:space:]]*\"/s/\"\(.*\)\"/\"$installusername\"/" ./flake.nix
+sed -i 's/username\s*=\s*"\([^"]*\)"/username = "'"$installusername"'"/' ./flake.nix
 
 echo "-----"
 
 echo "$NOTE Generating The Hardware Configuration"
-sudo nixos-generate-config --show-hardware-config > ./hosts/$hostName/hardware.nix
+attempts=0
+max_attempts=3
+hardware_file="./hosts/$hostName/hardware.nix"
+
+while [ $attempts -lt $max_attempts ]; do
+  sudo nixos-generate-config --show-hardware-config > "$hardware_file" 2>/dev/null
+
+  if [ -f "$hardware_file" ]; then
+    echo "${OK} Hardware configuration successfully generated."
+    break
+  else
+    echo "${WARN} Failed to generate hardware configuration. Attempt $(($attempts + 1)) of $max_attempts."
+    attempts=$(($attempts + 1))
+
+    # Exit if this was the last attempt
+    if [ $attempts -eq $max_attempts ]; then
+      echo "${ERROR} Unable to generate hardware configuration after $max_attempts attempts."
+      exit 1
+    fi
+  fi
+done
 
 echo "-----"
 
 echo "$NOTE Setting Required Nix Settings Then Going To Install"
-NIX_CONFIG="experimental-features = nix-command flakes"
+git config --global user.name "installer"
+git config --global user.email "installer@gmail.com"
+git add .
+sed -i 's/host\s*=\s*"\([^"]*\)"/host = "'"$hostName"'"/' ./flake.nix
 
 echo "-----"
 printf "\n%.0s" {1..2}
@@ -134,7 +156,10 @@ echo "-----"
 echo "$ERROR YES!!! YOU read it right.. you staring too much at your monitor ha ha... joke :)......"
 printf "\n%.0s" {1..2}
 
-sudo nixos-rebuild switch --flake ~/NixOS-Hyprland/#${hostName}
+# Set the Nix configuration for experimental features
+NIX_CONFIG="experimental-features = nix-command flakes"
+#sudo nix flake update
+sudo nixos-rebuild switch --flake ~/NixOS-Hyprland/#"${hostName}"
 
 echo "-----"
 printf "\n%.0s" {1..2}
